refactor: create own component for selectors & update useTransak hook
