+++ src/routes/embedded/wallet/buy/components/selector/index.ts
@@ -1,0 +1,2 @@
+export * from "./CurrencySelector";
+export * from "./PaymentSelector";
